Imported Upstream version 4.1.0-dfsg

+++ src/VBox/Runtime/common/misc/RTFileOpenF.cpp
@@ -1,4 +1,4 @@
-/* $Id: RTFileOpenF.cpp $ */
+/* $Id: RTFileOpenF.cpp 37596 2011-06-22 19:30:06Z vboxsync $ */
 /** @file
  * IPRT - RTFileOpenF.
  */
@@ -32,7 +32,7 @@
 #include "internal/iprt.h"
 
 
-RTR3DECL(int) RTFileOpenF(PRTFILE pFile, uint32_t fOpen, const char *pszFilenameFmt, ...)
+RTR3DECL(int) RTFileOpenF(PRTFILE pFile, uint64_t fOpen, const char *pszFilenameFmt, ...)
 {
     va_list va;
     va_start(va, pszFilenameFmt);
